Factor out internal ds_file functions.
