+++ dataswarm/src/manager/ds_manager.c
@@ -36,7 +36,6 @@ See the file COPYING for details.
 #include "random.h"
 #include "process.h"
 #include "path.h"
-#include "md5.h"
 #include "url_encode.h"
 #include "jx_print.h"
 #include "jx_parse.h"
@@ -136,9 +135,6 @@ int ds_option_scheduler = DS_SCHEDULE_TIME;
 
 /* default timeout for slow workers to come back to the pool */
 double ds_option_blocklist_slow_workers_timeout = 900;
-
-/* Internal use: when the worker uses the client library, do not recompute cached names. */
-int ds_hack_do_not_compute_cached_name = 0;
 
 /* time threshold to check when tasks are larger than connected workers */
 static timestamp_t interval_check_for_large_tasks = 180000000; // 3 minutes in usecs
@@ -1493,71 +1489,6 @@ static ds_result_code_t get_file_or_directory( struct ds_manager *q, struct ds_w
 	}
 
 	return result;
-}
-
-/*
-For a given task and file, generate the name under which the file
-should be stored in the remote cache directory.
-
-The basic strategy is to construct a name that is unique to the
-namespace from where the file is drawn, so that tasks sharing
-the same input file can share the same copy.
-
-In the common case of files, the cached name is based on the
-hash of the local path, with the basename of the local path
-included simply to assist with debugging.
-
-In each of the other file types, a similar approach is taken,
-including a hash and a name where one is known, or another
-unique identifier where no name is available.
-*/
-
-char *make_cached_name( const struct ds_file *f )
-{
-	static unsigned int file_count = 0;
-	file_count++;
-
-	/* Default of payload is remote name (needed only for directories) */
-	char *payload = f->payload ? f->payload : f->remote_name;
-
-	unsigned char digest[MD5_DIGEST_LENGTH];
-	char payload_enc[PATH_MAX];
-
-	if(f->type == DS_BUFFER) {
-		//dummy digest for buffers
-		md5_buffer("buffer", 6, digest);
-	} else {
-		md5_buffer(payload,strlen(payload),digest);
-		url_encode(path_basename(payload), payload_enc, PATH_MAX);
-	}
-
-	/* 0 for cache files, file_count for non-cache files. With this, non-cache
-	 * files cannot be shared among tasks, and can be safely deleted once a
-	 * task finishes. */
-	unsigned int cache_file_id = 0;
-	if(!(f->flags & DS_CACHE)) {
-		cache_file_id = file_count;
-	}
-
-	switch(f->type) {
-		case DS_FILE:
-		case DS_DIRECTORY:
-			return string_format("file-%d-%s-%s", cache_file_id, md5_string(digest), payload_enc);
-			break;
-		case DS_FILE_PIECE:
-			return string_format("piece-%d-%s-%s-%lld-%lld",cache_file_id, md5_string(digest),payload_enc,(long long)f->offset,(long long)f->piece_length);
-			break;
-		case DS_REMOTECMD:
-			return string_format("cmd-%d-%s", cache_file_id, md5_string(digest));
-			break;
-		case DS_URL:
-			return string_format("url-%d-%s", cache_file_id, md5_string(digest));
-			break;
-		case DS_BUFFER:
-		default:
-			return string_format("buffer-%d-%s", cache_file_id, md5_string(digest));
-			break;
-	}
 }
 
 /*
@@ -5093,37 +5024,6 @@ static void delete_feature(struct ds_task *t, const char *name)
 	list_cursor_destroy(c);
 }
 
-struct ds_file *ds_file_create(const char *payload, const char *remote_name, ds_file_t type, ds_file_flags_t flags)
-{
-	struct ds_file *f;
-
-	f = malloc(sizeof(*f));
-	if(!f) {
-		debug(D_NOTICE, "Cannot allocate memory for file %s.\n", remote_name);
-		return NULL;
-	}
-
-	memset(f, 0, sizeof(*f));
-
-	f->remote_name = xxstrdup(remote_name);
-	f->type = type;
-	f->flags = flags;
-
-	/* DS_BUFFER needs to set these after the current function returns */
-	if(payload) {
-		f->payload = xxstrdup(payload);
-		f->length  = strlen(payload);
-	}
-
-	if(ds_hack_do_not_compute_cached_name) {
-  		f->cached_name = xxstrdup(f->payload);
-	} else {
-		f->cached_name = make_cached_name(f);
-	}
-	
-	return f;
-}
-
 int ds_task_specify_url(struct ds_task *t, const char *file_url, const char *remote_name, ds_file_type_t type, ds_file_flags_t flags)
 {
 	struct list *files;
@@ -5490,16 +5390,6 @@ void ds_task_specify_monitor_output(struct ds_task *t, const char *monitor_outpu
 	}
 
 	t->monitor_output_directory = xxstrdup(monitor_output_directory);
-}
-
-void ds_file_delete(struct ds_file *tf) {
-	if(tf->payload)
-		free(tf->payload);
-	if(tf->remote_name)
-		free(tf->remote_name);
-	if(tf->cached_name)
-		free(tf->cached_name);
-	free(tf);
 }
 
 void ds_invalidate_cached_file(struct ds_manager *q, const char *local_name, ds_file_t type) {
